# Optimize repeated pointer accesses in op/base reduction functions