Hi, this PR changes some repeated pointer accesses in the reduction function range checks, so that they are only performed (the pointer accesses) once. (Hope I didn't miss any, are there any reduction functions in another file?)

I wouldn't expect this change to make a difference, and I probably would expect the compiler to optimize the original code on its own, but I have experimentally observed increased performance with this change. Might well not be the case for a different system or compiler, but the changes shouldn't really have any downside. 

Allreduce performance (latency) before patch
`10 runs: Average 6374.9 us, Stddev 30.2834 us, Min 6327.88 us, Max 6420.74 us`

Allreduce performance (latency) after patch
`10 runs: Average 5776.3 us, Stddev 41.5984 us, Min 5705.17 us, Max 5828.30 us`

Setup
```
osu_allreduce_cc @ 1M (data-altering osu_allreduce variant)
coll: libnbc,basic
pml: ob1
btl: vader,self (xpmem)
bind-to: core
map-to: slot

CentOS 8, GCC 8.5.0, aarch64 (single node)
--enable-debug not specified during config
```

Disclaimer: Tested on v5.0.0rc2

Signed-off-by: George Katevenis <gkatev@ics.forth.gr>