@jsquyres 2021-12-01
I'm a little surprised that we left that pointer-dereference assignment in all those loops, but I guess it just got missed.

I double checked MPI-4.0 to make sure that the len argument isn't an INOUT, and it doesn't appear to be.  MPI-4.0 p235:13-16 says:

> The len argument ... is passed in as a reference for reasons of compatibility with Fortran.

@bosilca Can you verify?