@ggouaillardet 2021-12-01
@jsquyres the Git commit checker did not run (even if you issue the `bot` command`) because this is a new contributor (!), so I had to manually hit a 'Approve and Run' button (iirc) in the web interface.