@ggouaillardet 2021-12-01
@gkatev thanks for the patch!

Out of curiosity, did you make any performance measurements (with Open MPI compiled for production - e.g. `-O2` or more aggressive)?
Though the patch looks correct at first glance, I would naively suspect the compiler optimize these pointer dereferences, ending up generating the same code with or without this patch.