@devreal 2021-12-01
This might be an aliasing issue: the compiler cannot assume that `count` isn't aliased by another pointer, unless it's marked as `restrict`. Hence the repeated load from memory.