@gkatev 2021-12-01
@ggouaillardet I think I am not seeing a difference with O2/O3, but am I doing it correctly like this?
`CFLAGS=-O2 make -C ompi`
Or is there a config-time or otherwise way to tune the optimization level?