@awlauria 2021-12-01
@gkatev I believe you'll need to pass in CFLAGS at configure time for it to have any effect.

you can check with `make -V2` to see it got passed in.