@awlauria 2021-12-01
FWIW in past projects I have seen these little optimizations making a difference. 

Compilers should be smart enough, but can't always rely on that.