@bosilca 2021-12-01
Check the assembly at [godbolt](https://godbolt.org/z/vb7rjx98c) for 4 different versions of the sum int32_t. Depending on the compiler, the generated code is slightly different. clang seems to assume the pointer to count is restrict and always generates the same optimized code, while gcc is taking a very cautious approach. In general the difference is reloading the *count from the cache instead of working on it from a register.

Conclusion, to remove the compiler from the equation and always be sure the code is as optimal as possible either we add restrict or we do the trick proposed here (temporary variable).