@bosilca 2021-12-01
However, I think I would rather have the code with the following pattern:

```c
void
ompi_op_base_2buff_sum_int32_t(const void *in, void *out, int *count)
  {
      int i;
      int32_t *a = (int32_t *) in;
      int32_t *b = (int32_t *) out;
      for (i = *count; i >= 0; i--) {
          *(b++) += *(a++);
      }
  }
```