@jsquyres 2021-12-01
@bosilca Just curious: why keep it incrementing through a pointer dereference?