@bosilca 2021-12-01
Not sure I understand the question. Are we talking about count or about the other pointers ?