@jsquyres 2021-12-01
@bosilca You're right... my question was non-sensical.  Disregard!