@devreal 2021-12-02
I thought the point of @bosilca 's comment was to walk backwards (`i--`) instead of forward (`i++`), which may be more cache efficient (assuming that an application wrote the last bytes last and will access again from the first bytes)