@bosilca 2021-12-02
I created #9719 as an alternative to this. While digging in the code I also noticed we are not using the 3buff MPI_Op so I took the opportunity to remove all support.