@gkatev 2021-12-02
FYI, for curiosity's sake I checked with `make V=2`, and it appears that all commands involving `op_base_functions.c` do contain `-O3`. (by default, without explicitely asking for it during configuration/compilation) -- and therefore all my tests here should have contained it.