@ggouaillardet 2021-12-03
@devreal you are absolutely right about pointer aliasing! and that makes the patch necessary to improve performance when the compiler chooses to generate correct code.