@jsquyres 2022-01-23
We merged #9719 and #9867 proposes removing the 3-buf variants.  Should this PR now be closed?